Make Conway-specific queries only available in Conway
