+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/TreeDiff.hs
@@ -167,3 +167,5 @@ instance
   , ToExpr (PredicateFailure (EraRule "CERTS" era))
   ) =>
   ToExpr (ConwayLedgerPredFailure era)
+
+instance ToExpr (Constitution era)
