+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/TreeDiff.hs
@@ -54,8 +54,6 @@ instance
   ) =>
   ToExpr (ShelleyGovState era)
 
-instance ToExpr (Constitution era)
-
 -- TxCert
 instance ToExpr (GenesisDelegCert c)
 instance ToExpr MIRPot
