+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.9.0.0
 
+* Remove `getDRepDistr`, `getConstitution` and `getCommitteeMembers` from `EraGov` #4033
 * Add `refScriptSize` parameter to `getMinFeeTx`, `setMinFeeTx` and `estimateMinFeeTx`
 * Add `hoistGovRelation`, `withGovActionParent` and `GovRelation`
 * Remove `PrevGovActionIds`
